10.19 2:34第十次提交 完成10.17修改意见

+++ src/main/java/com/yulaw/ccbapi/model/dao/ChannelAndVideoMapper.java
@@ -7,9 +7,9 @@ import java.util.List;
 
 @Repository
 public interface ChannelAndVideoMapper {
-    int deleteByPrimaryKey(Integer id);
+    int deleteByPrimaryKey(Long id);
 
-    ChannelAndVideo selectByPrimaryKey(Integer id);
+    ChannelAndVideo selectByPrimaryKey(Long id);
 
     int updateByPrimaryKeySelective(ChannelAndVideo record);
